2-Player Functionality
It's not done yet, but I added the ability to play the game with 2
players. It's not done yet, but it does track player scores. Right now,
a player gets 1 point for a correct guess. We may want to give a bonus
for guessing a letter that appears more than once